Refactor: directory name Uppercase 변경
